[Refactor]ElementsTileコンポーネント実装
